# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Bus Pickup API",
    "version": "1.0.0",
    "description": "API REST pour l'application de ramassage par bus d'entreprise",
    "contact": {
      "name": "API Support",
      "email": "support@buspickup.com"
    }
  },
  "servers": [
    {
      "url": "https://buspickup.vercel.app",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
  "tags": [
    {
      "name": "Rides",
      "description": "Gestion des demandes de ramassage"
    },
    {
      "name": "Drivers",
      "description": "Gestion des chauffeurs"
    },
    {
      "name": "Passengers",
      "description": "Gestion des passagers"
    }
  ],
  "paths": {
    "/api/v1/rides": {
      "get": {
        "summary": "Récupérer toutes les demandes de ramassage",
        "tags": ["Rides"],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["pending", "accepted", "completed", "cancelled"]
            },
            "description": "Filtrer par statut"
          },
          {
            "name": "passenger_id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filtrer par ID passager"
          },
          {
            "name": "driver_id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filtrer par ID chauffeur"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des demandes de ramassage",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/RideRequest"
                      }
                    },
                    "count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Créer une nouvelle demande de ramassage",
        "tags": ["Rides"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["pickup_lat", "pickup_lng", "pickup_address"],
                "properties": {
                  "pickup_lat": {
                    "type": "number",
                    "example": 48.8566
                  },
                  "pickup_lng": {
                    "type": "number",
                    "example": 2.3522
                  },
                  "pickup_address": {
                    "type": "string",
                    "example": "123 Rue de la Paix, Paris"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Demande créée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/RideRequest"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requête invalide"
          },
          "401": {
            "description": "Non autorisé"
          }
        }
      }
    },
    "/api/v1/rides/{id}": {
      "get": {
        "summary": "Récupérer une demande spécifique",
        "tags": ["Rides"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Détails de la demande",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/RideRequest"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Demande non trouvée"
          }
        }
      },
      "patch": {
        "summary": "Mettre à jour une demande",
        "tags": ["Rides"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": ["pending", "accepted", "completed", "cancelled"]
                  },
                  "driver_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Demande mise à jour"
          },
          "404": {
            "description": "Demande non trouvée"
          }
        }
      },
      "delete": {
        "summary": "Annuler une demande",
        "tags": ["Rides"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Demande annulée"
          },
          "404": {
            "description": "Demande non trouvée"
          }
        }
      }
    },
    "/api/v1/drivers": {
      "get": {
        "summary": "Récupérer tous les chauffeurs",
        "tags": ["Drivers"],
        "parameters": [
          {
            "name": "active",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "Filtrer par statut actif"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des chauffeurs"
          }
        }
      }
    },
    "/api/v1/passengers": {
      "get": {
        "summary": "Récupérer tous les passagers",
        "tags": ["Passengers"],
        "responses": {
          "200": {
            "description": "Liste des passagers"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RideRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "passenger_id": {
            "type": "string",
            "format": "uuid"
          },
          "driver_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "pickup_lat": {
            "type": "number"
          },
          "pickup_lng": {
            "type": "number"
          },
          "pickup_address": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "accepted", "completed", "cancelled"]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "accepted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      }
    }
  }
}